Challenge 기여를 한 transaction 으로 묶는다.
