+++ app-server/subprojects/bounded_context/accessibility/application/build.gradle.kts
@@ -1,4 +1,5 @@
 dependencies {
     implementation(projects.boundedContext.user.application)
     implementation(projects.boundedContext.place.application)
+    implementation(projects.boundedContext.challenge.application)
 }
