+++ app-server/subprojects/bounded_context/accessibility/application/src/main/kotlin/club/staircrusher/accessibility/application/port/in/RegisterPlaceAccessibilityUseCase.kt
@@ -4,6 +4,9 @@ import club.staircrusher.accessibility.application.port.`in`.result.GetAccessibi
 import club.staircrusher.accessibility.application.port.`in`.result.RegisterPlaceAccessibilityResult
 import club.staircrusher.accessibility.application.port.out.persistence.PlaceAccessibilityCommentRepository
 import club.staircrusher.accessibility.application.port.out.persistence.PlaceAccessibilityRepository
+import club.staircrusher.challenge.application.port.`in`.ChallengeService
+import club.staircrusher.challenge.domain.model.ChallengeAddress
+import club.staircrusher.challenge.domain.model.ChallengeContribution
 import club.staircrusher.stdlib.di.annotation.Component
 import club.staircrusher.stdlib.persistence.TransactionManager
 
@@ -11,14 +14,46 @@ import club.staircrusher.stdlib.persistence.TransactionManager
 class RegisterPlaceAccessibilityUseCase(
     private val transactionManager: TransactionManager,
     private val accessibilityApplicationService: AccessibilityApplicationService,
+    private val challengeService: ChallengeService
 ) {
+    data class RegisterPlaceAccessibilityUseCaseResult(
+        val registerPlaceAccessibilityResult: RegisterPlaceAccessibilityResult,
+        val getAccessibilityResult: GetAccessibilityResult,
+        val challengeContributions: List<ChallengeContribution>
+    )
+
     fun handle(
         userId: String?,
         createPlaceAccessibilityParams: PlaceAccessibilityRepository.CreateParams,
         createPlaceAccessibilityCommentParams: PlaceAccessibilityCommentRepository.CreateParams?,
-    ) : Pair<RegisterPlaceAccessibilityResult, GetAccessibilityResult> = transactionManager.doInTransaction {
-        val registerResult = accessibilityApplicationService.doRegisterPlaceAccessibility(createPlaceAccessibilityParams, createPlaceAccessibilityCommentParams)
-        val getAccessibilityResult = accessibilityApplicationService.doGetAccessibility(createPlaceAccessibilityParams.placeId, userId)
-        Pair(registerResult, getAccessibilityResult)
+    ): RegisterPlaceAccessibilityUseCaseResult = transactionManager.doInTransaction {
+        val registerResult = accessibilityApplicationService.doRegisterPlaceAccessibility(
+            createPlaceAccessibilityParams,
+            createPlaceAccessibilityCommentParams
+        )
+        val getAccessibilityResult =
+            accessibilityApplicationService.doGetAccessibility(createPlaceAccessibilityParams.placeId, userId)
+        val challengeContributions = userId?.let { userId ->
+            challengeService.contributeToSatisfiedChallenges(
+                userId = userId,
+                contribution = ChallengeService.Contribution.PlaceAccessibility(
+                    placeAccessibilityId = registerResult.placeAccessibility.id,
+                    placeAccessibilityAddress = registerResult.place.address.let {
+                        ChallengeAddress(
+                            siDo = it.siDo,
+                            siGunGu = it.siGunGu,
+                            eupMyeonDong = it.eupMyeonDong,
+                            li = it.li,
+                            roadName = it.roadName
+                        )
+                    }
+                )
+            )
+        } ?: listOf()
+        return@doInTransaction RegisterPlaceAccessibilityUseCaseResult(
+            registerPlaceAccessibilityResult = registerResult,
+            getAccessibilityResult = getAccessibilityResult,
+            challengeContributions = challengeContributions
+        )
     }
 }
